// src/interfaces/iPersonalDataResponse.ts
export interface iPersonalDataResponse {
  isClient: boolean;
  clientType?: string;
  firstName: string;
  lastName: string;
  fullName?: string;
  documentNumber: string;
  documentType: string;
  email: string;
  cellPhone: string;
  phone: string;
  birthDay?: string;
  address: string;
  residenceCity?: string;
}